% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{add_iso_code}
\alias{add_iso_code}
\title{Add Country ISO Code column}
\usage{
add_iso_code(item, country_code)
}
\arguments{
\item{item}{country subdivision \code{data.frame} element}

\item{country_code}{2 letter iso code to assign}
}
\value{
Returns the inserted \code{data.frame} with the isocode as column attached.
}
\description{
Function to add country 2 letter isocode column in subidivision \code{data.frame} item
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
