% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{api_response}
\alias{api_response}
\title{Default HTTPS status error messages}
\source{
For more information about the Holiday API go to the \href{https://holidayapi.com/docs}{Holiday API documentation page}
}
\usage{
api_response(status)
}
\arguments{
\item{status}{HTTPS returned status code when doing a GET REST API call to any https://holidayapi.com endpoint.}
}
\value{
Returns a \code{character} with a predefined success/error linked to the status code.
}
\description{
Function which returns predefined default success/error messages as documented at \href{https://holidayapi.com/docs}{}.
}
\examples{
\dontrun{
api_response(200)
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
