% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{as.data.frame.available_languages}
\alias{as.data.frame.available_languages}
\title{Creates data.frame}
\usage{
\method{as.data.frame}{available_languages}(x)
}
\arguments{
\item{x}{Must be a \code{available_languages} class object}
}
\value{
Returns a \code{data.frame}.
}
\description{
Creates data.frame
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
