% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{as.data.frame.country_holidays}
\alias{as.data.frame.country_holidays}
\title{Creates data.frame}
\usage{
\method{as.data.frame}{country_holidays}(x)
}
\arguments{
\item{x}{Must be a \code{available_languages} class object}
}
\value{
Returns a summary of a \code{available_languages} class object.
}
\description{
Creates data.frame
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
