% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Creates data.frame}
\usage{
as.data.frame(x)
}
\arguments{
\item{x}{Must be a \code{country_holidays} object.}
}
\value{
Returns a base \code{data.frame}.
}
\description{
Creates a \code{data.frame} from a \code{country_holidays}, \code{available_countries} or \code{available_languages} class object.
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
