% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{as_available_countries}
\alias{as_available_countries}
\title{Convert to \code{available_countries} class}
\usage{
as_available_countries(request_url, key, call, ...)
}
\arguments{
\item{request_url}{URL with which the RESTful GET API call has ben made.}

\item{key}{The users API key obtained when registering at \href{https://holidayapi.com/signup}{}.}

\item{call}{JSON Response from the RESTful GET API call formatted back to a \code{list} element.}

\item{...}{Optional parameters.}
}
\value{
Returns a \code{available_countries} object class
}
\description{
Convert call response to an internal \code{available_countries} class object, formatting the response to ease tha holidays analysis.
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
