% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{as_country_holidays}
\alias{as_country_holidays}
\title{Convert to \code{country_holidays} class}
\usage{
as_country_holidays(request_url, key, country, year, call, ...)
}
\arguments{
\item{request_url}{URL with which the RESTful GET API call has ben made.}

\item{key}{The users API key obtained when registering at \href{https://holidayapi.com/signup}{}.}

\item{country}{Country code in \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha 2}
or \href{https://en.wikipedia.org/wiki/ISO_3166-2}{ISO 3166-2} format. For multiple countries,
pass a vector - it is limited to 10 countries. It is a required parameter in \code{character} format.}

\item{year}{Year in \href{https://en.wikipedia.org/wiki/ISO_8601#Years}{ISO 8601} format.
It is a required parameter in \code{numeric} format.}

\item{call}{JSON Response from the RESTful GET API call formatted back to a \code{list} element.}

\item{...}{Optional parameters.}
}
\value{
Returns a \code{country_holidays} object class
}
\description{
Convert call response to an internal \code{country_holidays} class object, formatting the response to ease tha holidays analysis.
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
