% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{default_version}
\alias{default_version}
\title{Returns API default version}
\usage{
default_version()
}
\value{
Returns the package API version
}
\description{
Allows to easily adjust the versioning within the package
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
