% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_holidays.R
\name{get_countries}
\alias{get_countries}
\title{Retrieve available countries}
\source{
For more information about the Holiday API go to the \href{https://holidayapi.com/docs}{Holiday API documentation page}
}
\usage{
get_countries(key = get_key())
}
\arguments{
\item{key}{The users API key obtained when registering at \href{https://holidayapi.com/signup}{}.
If the key has been saved to the environment before using the \code{save_key(key)} function,
it is not required to pass. It is a required parameter in \code{character} format.}
}
\value{
Returns a \code{available_countries} object containing the
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha 2} country code and the english short name.
Additionally, relevant information about the API request is shown.
}
\description{
Retrieve all 249 countries for which the Holiday API has holidays available.
Useful to pass on later on \code{to get_holidays()} function. By default the JSON format
 is selected because it contains additional request information.
}
\section{Warning}{

Depending on the subscription plan the scope of the API changes. At date 2020-04-01 the
free plan is limited to historical data for the previous year with 10.000 monthly requests in english
language format only. Non-commercial use only. For more, updated information go \href{https://holidayapi.com/}{here}.
}

\examples{
\dontrun{
save_key(key="__YOUR_API_KEY__")
get_countries()
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
