% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_holidays.R
\name{get_holidays}
\alias{get_holidays}
\title{Retrieve holidays}
\source{
For more information about the Holiday API go to the \href{https://holidayapi.com/docs}{Holiday API documentation page}
}
\usage{
get_holidays(country, year, key = get_key(), ...)
}
\arguments{
\item{country}{Country code in \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha 2}
or \href{https://en.wikipedia.org/wiki/ISO_3166-2}{ISO 3166-2} format. For multiple countries,
pass a vector - it is limited to 10 countries. It is a required parameter in \code{character} format.}

\item{year}{Year in \href{https://en.wikipedia.org/wiki/ISO_8601#Years}{ISO 8601} format.
It is a required parameter in \code{numeric} format.}

\item{key}{The users API key obtained when registering at \href{https://holidayapi.com/signup}{}.
If the key has been saved to the environment before using the \code{save_key(key)} function,
it is not required to pass. It is a required parameter in \code{character} format.}

\item{...}{Ellipsis to pass optional arguments into the function such as:
\itemize{
\item \code{month} - 1 or 2 digit month (1-12). It is an optional parameter.
\item \code{day} - 1 or 2 digit day (1-31 depending on the month). It is an optional parameter format.
\item \code{previous} - 1 or 2 digit day (1-31 depending on the month). It is an optional parameter format.
\item \code{upcoming} - Return upcoming holidays based on the date the API is called. It is an optional parameter format.
\item \code{public} - Return only \href{https://en.wikipedia.org/wiki/Public_holiday}{public holidays}.
It is an optional parameter format in boolean format.
\item \code{language} - By default English language, optionally other languages in ISO 639-1 language
#'  format can be passed (with exceptions). It is an optional parameter format in \code{character} format.
}}
}
\value{
Returns a \code{country_holidays} object containing the dates, names of holidays, if they
are public or not. Additionally, relevant information about the API request is shown.
}
\description{
Retrieve a list of holidays from a given country in a given year.
By default the JSON format is selected because it contains additional request information.
}
\section{Warning}{

Depending on the subscription plan the scope of the API changes. At date 2019-07-01 the free plan
is limited to historical data for 2018 with 10.000 monthly requests in english language format only.
 Non-commercial use only. For more, updated information go \href{https://holidayapi.com/}{here}.
}

\examples{
\dontrun{
get_holidays(country="ES", year=2018)
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
