% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{get_key}
\alias{get_key}
\title{Retrieves the key from users environment}
\usage{
get_key()
}
\value{
Returns the key from the users environment.
}
\description{
Retrieves the key assigned to the variable HOLIDAYAPI_PAT in the users environment.
The key needs to be explicitly saved with the \code{save_key(key)} function.
}
\examples{
\dontrun{get_key()}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
