% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_holidays.R
\name{get_languages}
\alias{get_languages}
\title{Retrieve available languages}
\source{
For more information about the Holiday API go to the \href{https://holidayapi.com/docs}{Holiday API documentation page}
}
\usage{
get_languages(key = get_key())
}
\arguments{
\item{key}{The users API key obtained when registering at \href{https://holidayapi.com/signup}{}.
If the key has been saved to the environment before using the \code{save_key(key)} function, it is not required to pass.
It is a required parameter in \code{character} format.}
}
\value{
Returns a \code{available_languages} object containing the
\href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 format (with exceptions)} language and its name.
Additionally, relevant information about the API request is shown.
}
\description{
Retrieve all 100 languages for which the Holiday API has holidays translation available.
By default returns english. Useful to pass on later on \code{to get_holidays()} function.
}
\section{Warning}{

Depending on the subscription plan the scope of the API changes. At date 2020-04-01 the free plan
is limited to historical data for the previous year with 10.000 monthly requests in english language format only.
Non-commercial use only. For more, updated information go \href{https://holidayapi.com/}{here}.
}

\examples{
\dontrun{
save_key(key="__YOUR_API_KEY__")
get_languages()
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
