% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{nonvalues2NA}
\alias{nonvalues2NA}
\title{Converts non-usable values to empty character string}
\usage{
nonvalues2NA(item)
}
\arguments{
\item{item}{Logical FALSE, empty, NULL or NA element}
}
\value{
Returns an empty \code{character} string.
}
\description{
Converts FALSE, character(0), NA, list(), NULL to an empty  \code{""} \code{character} string
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
