% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{print.available_countries}
\alias{print.available_countries}
\title{Print a \code{available_countries} class object}
\usage{
\method{print}{available_countries}(x)
}
\arguments{
\item{x}{Must be a \code{available_countries} class object.}
}
\value{
Prints a \code{available_countries} class object.
}
\description{
Print a \code{available_countries} class object
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
