% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{print.available_languages}
\alias{print.available_languages}
\title{Print a \code{available_languages} class}
\usage{
\method{print}{available_languages}(x)
}
\arguments{
\item{x}{Must be a \code{available_languages} class object.}
}
\value{
Prints a \code{available_languages} class object.
}
\description{
Print a \code{available_languages} class
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
