% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{print}
\alias{print}
\title{Print Holiday API classes}
\usage{
print(x)
}
\arguments{
\item{x}{Can be a \code{country_holidays}, \code{available_countries} or \code{available_languages} object.}
}
\value{
Prints a Holiday API class.
}
\description{
Print Holiday API classes
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
