% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_holidays.R
\name{save_key}
\alias{save_key}
\title{Saves the key for future functions}
\usage{
save_key(key)
}
\arguments{
\item{key}{The users API key obtained when registering at \href{https://holidayapi.com/signup}{}.}
}
\value{
Saves the key in the users environment as HOLIDAYAPI_PAT - it does not return any object.
}
\description{
Saves the key in the users environment as HOLIDAYAPI_PAT. It has the advantage that it is not necessary to explicitly publish the key in the users code. Just do it ones and you're set. To update the key kust save again and it will overwrite the old key.
}
\examples{
\dontrun{
save_key(key="__YOUR_API_KEY__")
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
