% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{summary.available_countries}
\alias{summary.available_countries}
\title{Summary of a \code{available_countries} class}
\usage{
\method{summary}{available_countries}(x)
}
\arguments{
\item{x}{Must be a \code{available_countries} class object.}
}
\value{
Returns a summary of a \code{available_countries} class object.
}
\description{
Summary of a \code{available_countries} class
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
