% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{summary.available_languages}
\alias{summary.available_languages}
\title{Summary of a \code{available_languages} class}
\usage{
\method{summary}{available_languages}(x)
}
\arguments{
\item{x}{Must be a \code{available_languages} class object.}
}
\value{
Returns a summary of a \code{available_languages} class object.
}
\description{
Summary of a \code{available_languages} class
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
