% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{summary.country_holidays}
\alias{summary.country_holidays}
\title{Summary of a \code{country_holidays} class}
\usage{
\method{summary}{country_holidays}(x)
}
\arguments{
\item{x}{Must be a \code{country_holidays} class object.}
}
\value{
Returns a summary of a \code{country_holidays} class object.
}
\description{
Summary of a \code{country_holidays} class
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
