% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{summary}
\alias{summary}
\title{Summary of Holiday API classes}
\usage{
summary(x)
}
\arguments{
\item{x}{Can be a \code{country_holidays}, \code{available_countries} or \code{available_languages} object}
}
\value{
Returns a summary of a Holiday API class
}
\description{
Summary of Holiday API classes
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
