% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{unnest_df}
\alias{unnest_df}
\title{Unnest a nested \code{data.frame}}
\usage{
unnest_df(x)
}
\arguments{
\item{x}{A nested \code{data.frame} object}
}
\value{
Returns a flattened \code{data.frame}.
}
\description{
Unnests/flattens a nested \code{data.frame}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
