% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods_other.R
\name{print.country_holidays}
\alias{print.country_holidays}
\title{Print a \code{country_holidays} class}
\usage{
\method{print}{country_holidays}(x)
}
\arguments{
\item{x}{Must be a \code{country_holidays} class object.}
}
\value{
Prints a \code{country_holidays} class object.
}
\description{
Print a \code{country_holidays} class
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
